navigation: Unset the narrative text when the icon changes
Navigation narrative commonly comes only at specific intervals and at least
OSMand does update the direction icon and countdown-distance way earlier.

This can lead to situations when the old narrative is displayed to a
mismatched direction icon.

So after the icon changes, assume the narrative is not applicable anymore.

The flag-update in the code also happens before possibly updating the
narrative, so if a new narrative text is send within the same update,
this will happen just normally.