navigation: align the flag from the top of the screen
The current navigation layout starts from the center of the screen for
flag, narrative and distance.

With the current offsets, this puts the flag in coordinates

    (80,20 - 159,99)

on a 240x240 PineTime display, leaving 20px unused space above it.
If the display of the device is bigger, that unused space would be even
bigger.

With space for navigation information at a premium, lets start a bit of
reorganization to make more room, starting with aligning the flag from
the top and only use 2px of empty space above it.